Remove useless code (#50)
